Update db 1.5

+++ BMI_Tracker/src/main/java/com/fu/bmi_tracker/controller/MealLogController.java
@@ -5,14 +5,12 @@
 package com.fu.bmi_tracker.controller;
 
 import com.fu.bmi_tracker.exceptions.ErrorMessage;
-import com.fu.bmi_tracker.model.entities.MealLog;
 import com.fu.bmi_tracker.model.entities.CustomAccountDetailsImpl;
-import com.fu.bmi_tracker.model.entities.Food;
+import com.fu.bmi_tracker.model.entities.DailyRecord;
 import com.fu.bmi_tracker.model.entities.MealLog;
 import com.fu.bmi_tracker.model.entities.Member;
-import com.fu.bmi_tracker.model.enums.EMealType;
 import com.fu.bmi_tracker.payload.request.CreateMealLogRequest;
-import com.fu.bmi_tracker.services.FoodService;
+import com.fu.bmi_tracker.services.DailyRecordService;
 import com.fu.bmi_tracker.services.MealLogService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
@@ -23,7 +21,6 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.util.Date;
-import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -36,6 +33,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import com.fu.bmi_tracker.services.MemberService;
 import jakarta.validation.Valid;
+import java.util.Optional;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -57,9 +55,9 @@ public class MealLogController {
     MemberService memberService;
 
     @Autowired
-    FoodService foodService;
+    DailyRecordService dailyRecordService;
 
-    @Operation(summary = "Retrieve meal list (Member)", description = "Get list meal of date from account id")
+    @Operation(summary = "Retrieve meal list (Member)", description = "Get list meal of date")
     @ApiResponses({
         @ApiResponse(responseCode = "200", content = {
             @Content(schema = @Schema(implementation = MealLog.class), mediaType = "application/json")}),
@@ -85,8 +83,16 @@ public class MealLogController {
         CustomAccountDetailsImpl principal = (CustomAccountDetailsImpl) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
         Member member = memberService.findByAccountID(principal.getId()).get();
 
-        // Get Meal log with date
-        Iterable<MealLog> mealLogs = mealLogService.findAllByDateOfMealAndMember_MemberID(dateOfMeal, member.getMemberID());
+        //Find record ID by memberID and Date
+        Optional<DailyRecord> record = dailyRecordService.findByMemberIDAndDate(member.getMemberID(), dateOfMeal);
+
+        // check record
+        if (!record.isPresent()) {
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
+
+        // Get Meal log with recordID
+        Iterable<MealLog> mealLogs = mealLogService.findByRecordID(record.get().getRecordID());
 
         // check meal empty
         if (!mealLogs.iterator().hasNext()) {
@@ -94,49 +100,6 @@ public class MealLogController {
         }
 
         return new ResponseEntity<>(mealLogs, HttpStatus.OK);
-    }
-
-    @Operation(summary = "Retrieve food list (MEMBER)", description = "Member retrieve food list of meal by mealtype and date")
-    @ApiResponses({
-        @ApiResponse(responseCode = "200", content = {
-            @Content(schema = @Schema(implementation = Food.class), mediaType = "application/json")}),
-        @ApiResponse(responseCode = "204", description = "There are no Tags", content = {
-            @Content(schema = @Schema())}),
-        @ApiResponse(responseCode = "500", content = {
-            @Content(schema = @Schema())})})
-    @GetMapping("/getFoodOfMeal")
-    @PreAuthorize("hasRole('MEMBER')")
-    public ResponseEntity<?> getFoodListByMealType(
-            @RequestParam String date,
-            @RequestParam EMealType mealType
-    ) {
-        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
-
-        LocalDate dateOfMeal;
-        // Validation date
-        try {
-            dateOfMeal = LocalDate.parse(date, formatter);
-        } catch (Exception e) {
-            ErrorMessage errorMessage = new ErrorMessage(HttpStatus.BAD_REQUEST.value(), new Date(), "Invalid date format. Please provide the date in the format yyyy-MM-dd.", "");
-            return new ResponseEntity<>(errorMessage, HttpStatus.BAD_REQUEST);
-        }
-
-        CustomAccountDetailsImpl principal = (CustomAccountDetailsImpl) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-
-        Member member = memberService.findByAccountID(principal.getId()).get();
-
-        // Get list food id 
-        Iterable<Integer> foodIDs = mealLogService.findFoodIDByDateOfMealAndMemberIDAndMealType(dateOfMeal,
-                member.getMemberID(), mealType);
-
-        if (!foodIDs.iterator().hasNext()) {
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
-        }
-
-        // Get food from list food id
-        Iterable<Food> foods = foodService.findByFoodIDIn((List<Integer>) foodIDs);
-
-        return new ResponseEntity<>(foods, HttpStatus.OK);
     }
 
     @Operation(summary = "Retrieve total calories meal of date", description = "Get total calories meal of date from account id")
@@ -165,21 +128,10 @@ public class MealLogController {
         CustomAccountDetailsImpl principal = (CustomAccountDetailsImpl) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
         Member member = memberService.findByAccountID(principal.getId()).get();
 
-        // Get Meal log with date
-        Iterable<MealLog> mealLogs = mealLogService.findAllByDateOfMealAndMember_MemberID(dateOfMeal, member.getMemberID());
+        //Find record ID by memberID and Date
+        Optional<DailyRecord> record = dailyRecordService.findByMemberIDAndDate(member.getMemberID(), dateOfMeal);
 
-        // check meal empty
-        if (!mealLogs.iterator().hasNext()) {
-            return new ResponseEntity<>(0, HttpStatus.NO_CONTENT);
-        }
-
-        // Calculate calories
-        int totalCalories = 0;
-        for (MealLog mealLog : mealLogs) {
-            totalCalories += mealLog.getCalories();
-        }
-
-        return new ResponseEntity<>(totalCalories, HttpStatus.OK);
+        return new ResponseEntity<>(record.get().getTotalCaloriesIn(), HttpStatus.OK);
     }
 
     @Operation(
@@ -198,6 +150,31 @@ public class MealLogController {
         // Create new object
         MealLog mealLog = new MealLog(mealLogRequest);
 
+        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
+        LocalDate dateOfMeal;
+
+        // Validation date 
+        try {
+            dateOfMeal = LocalDate.parse(mealLogRequest.getDateOfMeal(), formatter);
+        } catch (Exception e) {
+            ErrorMessage errorMessage = new ErrorMessage(HttpStatus.BAD_REQUEST.value(), new Date(), "Invalid date format. Please provide the date in the format yyyy-MM-dd.", "");
+            return new ResponseEntity<>(errorMessage, HttpStatus.BAD_REQUEST);
+        }
+
+        // Get Member id from acccount id context
+        CustomAccountDetailsImpl principal = (CustomAccountDetailsImpl) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        Member member = memberService.findByAccountID(principal.getId()).get();
+
+        DailyRecord dailyRecord = dailyRecordService.findByMemberIDAndDate(member.getMemberID(), dateOfMeal).get();
+
+        if (dailyRecord == null) {
+            dailyRecord = dailyRecordService.save(new DailyRecord(dateOfMeal, member.getMemberID()));
+
+        }
+
+        //set record ID
+        mealLog.setRecordID(dailyRecord.getRecordID());
+
         // Store to database
         MealLog mealLogSave = mealLogService.save(mealLog);
 
@@ -206,6 +183,13 @@ public class MealLogController {
             return new ResponseEntity<>("Failed to create new mealLog", HttpStatus.INTERNAL_SERVER_ERROR);
 
         }
+
+        // update calories in of daily record
+        int caloriesIn = dailyRecord.getTotalCaloriesIn() + mealLogSave.getCalories();
+
+        dailyRecord.setTotalCaloriesIn(caloriesIn);
+
+        dailyRecordService.save(dailyRecord);
 
         return new ResponseEntity<>(mealLogSave, HttpStatus.CREATED);
     }
@@ -221,7 +205,23 @@ public class MealLogController {
     @DeleteMapping("/delete")
     @PreAuthorize("hasRole('MEMBER')")
     public ResponseEntity<?> deleteMealLogById(@RequestParam int mealLogID) {
-        mealLogService.deleteById(mealLogID);
+        // Find meal log by meal logid
+        Optional<MealLog> mealLog = mealLogService.findById(mealLogID);
+        // check existed
+        if (mealLog.isPresent()) {
+            //find Daily record
+            Optional<DailyRecord> dailyRecord = dailyRecordService.findById(mealLog.get().getRecordID());
+            //Update coloriesIn
+            int coloriesIn = dailyRecord.get().getTotalCaloriesIn() - mealLog.get().getCalories();
+            dailyRecord.get().setTotalCaloriesIn(coloriesIn);
+
+            // delete meal log
+            mealLogService.deleteById(mealLogID);
+
+            //update dailyrecord
+            dailyRecordService.save(dailyRecord.get());
+        }
+
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 }
